fix(oxlint-config): Provide JS instead of TS for node_modules compatibility
