--- packages/oxlint-config/configs/index.ts
@@ -1,13 +1,0 @@
-import testEssentials from '../src/configs/test/essentials.ts';
-
-const test = {
-  essentials: testEssentials,
-};
-
-export { default as essentials } from '../src/configs/essentials.ts';
-export { default as jsdoc } from '../src/configs/jsdoc.ts';
-export { default as nextjs } from '../src/configs/nextjs.ts';
-export { default as node } from '../src/configs/node.ts';
-export { default as react } from '../src/configs/react.ts';
-export { default as typescript } from '../src/configs/typescript.ts';
-export { test };
