+++ packages/oxlint-config/oxlint.config.ts
@@ -1,12 +1,12 @@
 import { defineConfig } from 'oxlint';
-import essentials from './src/configs/essentials.ts';
-import typescript from './src/configs/typescript.ts';
+import essentials from './src/configs/essentials.js';
+import typescript from './src/configs/typescript.js';
 
 export default defineConfig({
   extends: [essentials, typescript],
   overrides: [
     {
-      files: ['./src/{configs,rules}/**/*.ts'],
+      files: ['./src/{configs,rules}/**/*.js'],
       rules: {
         'import/no-default-export': ['off'],
         'unicorn/filename-case': ['off'],
