+++ packages/oxlint-config/src/configs/nextjs.js
@@ -1,5 +1,5 @@
 import { defineConfig } from 'oxlint';
-import nextjsRuleset from '../rules/nextjs.ts';
+import nextjsRuleset from '../rules/nextjs.js';
 
 export default defineConfig({
   extends: [nextjsRuleset],
