+++ packages/oxlint-config/src/configs/react.js
@@ -1,6 +1,6 @@
 import { defineConfig } from 'oxlint';
-import jsxA11yRuleSet from '../rules/jsx-a11y.ts';
-import reactRuleSet from '../rules/react.ts';
+import jsxA11yRuleSet from '../rules/jsx-a11y.js';
+import reactRuleSet from '../rules/react.js';
 
 export default defineConfig({
   extends: [reactRuleSet, jsxA11yRuleSet],
