+++ packages/oxlint-config/src/configs/test/essentials.js
@@ -1,5 +1,5 @@
 import { defineConfig } from 'oxlint';
-import vitestRuleSet from '../../rules/vitest.ts';
+import vitestRuleSet from '../../rules/vitest.js';
 
 export default defineConfig({
   extends: [vitestRuleSet],
