+++ packages/oxlint-config/src/rules/eslint.js
