+++ packages/oxlint-config/src/rules/jest.js
