+++ packages/oxlint-config/src/rules/jsdoc.js
