+++ packages/oxlint-config/src/rules/jsx-a11y.js
