+++ packages/oxlint-config/src/rules/nextjs.js
