+++ packages/oxlint-config/src/rules/node.js
