+++ packages/oxlint-config/src/rules/oxc.js
