+++ packages/oxlint-config/src/rules/promise.js
