+++ packages/oxlint-config/src/rules/react.js
