+++ packages/oxlint-config/src/rules/typescript.js
