+++ packages/oxlint-config/src/rules/unicorn.js
