+++ packages/oxlint-config/src/rules/vitest.js
