+++ packages/oxlint-config/test/node/oxlint.config.ts
@@ -1,6 +1,6 @@
 import { defineConfig } from 'oxlint';
-import essentials from '../../src/configs/essentials.ts';
-import node from '../../src/configs/node.ts';
+import essentials from '../../src/configs/essentials.js';
+import node from '../../src/configs/node.js';
 
 export default defineConfig({
   extends: [essentials, node],
