+++ packages/oxlint-config/test/react/oxlint.config.ts
@@ -1,8 +1,8 @@
 import { defineConfig } from 'oxlint';
-import essentials from '../../src/configs/essentials.ts';
-import jsdoc from '../../src/configs/jsdoc.ts';
-import react from '../../src/configs/react.ts';
-import typescript from '../../src/configs/typescript.ts';
+import essentials from '../../src/configs/essentials.js';
+import jsdoc from '../../src/configs/jsdoc.js';
+import react from '../../src/configs/react.js';
+import typescript from '../../src/configs/typescript.js';
 
 export default defineConfig({
   extends: [essentials, typescript, jsdoc, react],
