+++ packages/oxlint-config/test/test/oxlint.config.ts
@@ -1,7 +1,7 @@
 import { defineConfig } from 'oxlint';
-import essentials from '../../src/configs/essentials.ts';
-import testEssentials from '../../src/configs/test/essentials.ts';
-import typescript from '../../src/configs/typescript.ts';
+import essentials from '../../src/configs/essentials.js';
+import testEssentials from '../../src/configs/test/essentials.js';
+import typescript from '../../src/configs/typescript.js';
 
 export default defineConfig({
   extends: [essentials, typescript, testEssentials],
