+++ packages/oxlint-config/test/typescript/oxlint.config.ts
@@ -1,6 +1,6 @@
 import { defineConfig } from 'oxlint';
-import essentials from '../../src/configs/essentials.ts';
-import typescript from '../../src/configs/typescript.ts';
+import essentials from '../../src/configs/essentials.js';
+import typescript from '../../src/configs/typescript.js';
 
 export default defineConfig({
   extends: [essentials, typescript],
