+++ packages/oxlint-config/tsconfig.json
@@ -5,7 +5,8 @@
     "moduleResolution": "NodeNext",
     "types": ["node", "vitest/globals"],
     "noEmit": true,
-    "allowImportingTsExtensions": true
+    "allowImportingTsExtensions": true,
+    "allowJs": true
   },
-  "include": ["./**/*.ts"]
+  "include": ["./**/*.ts", "configs/index.js"]
 }
